test(hooks): merge websocket tests into useAgentState.test and remove legacy test file
